Заменил на реквестпарам, убрал патч и делит, добавил ридонли, убрал костыль.

+++ src/main/java/web/dao/UserDAO.java
@@ -6,6 +6,6 @@ public interface UserDAO {
     List<User>getAllUsers ();
     User getUserById(long id);
     void addUser(User user);
-    void removeUser(long id);
+    void removeUser(int id);
     void updateUser(User user);
 }
